+++ resources/ReleaseNotes.md
@@ -6,5 +6,6 @@
 - Create `mag.meta.Mode` enumeration to capture sensor science mode
 - Move definition of time constants to shared utility file `mag.time.Constant`
 - Make `mag.process.Range` and `mag.process.SignedInteger` more flexible to custom variable names
+- Make sure converted values in `mag.process.SignedInteger` are returned as `double`
 - Fix [#10](https://github.com/ImperialCollegeLondon/MAG-Data-Visualization-Toolbox/issues/10)
 - Fix [#19](https://github.com/ImperialCollegeLondon/MAG-Data-Visualization-Toolbox/issues/19)